use following games in context state
